fix: 修改网络部分故障回调函数的故障码参数为4字节

+++ net/net_tha/inc/tha_fault_analyse.h
@@ -14,7 +14,7 @@
 
 #ifdef NET_PACK_USING_THA
 
-void tha_fault_event_detect_callback(uint8_t gunno, uint8_t code, uint32_t timestamp, uint8_t is_resume);
+void tha_fault_event_detect_callback(uint8_t gunno, uint32_t code, uint32_t timestamp, uint8_t is_resume);
 uint32_t tha_get_current_fault_set(uint8_t gunno);
 void tha_fault_detect_report(uint8_t gunno);
