+++ net_interface/inc/net_sal.h
@@ -125,7 +125,7 @@
 /****************************************************************/
 
 void app_nsal_realtime_process(uint8_t gunno);
-void app_nsal_system_fault_report(uint8_t gunno, uint8_t code, uint32_t timestamp, uint8_t is_resume);
+void app_nsal_system_fault_report(uint8_t gunno, uint32_t code, uint32_t timestamp, uint8_t is_resume);
 
 void app_nsal_message_init(uint8_t gunno);
 void app_nsal_state_charged(uint8_t gunno);
